+++ CHANGELOG.md
@@ -9,7 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
-- Add ValidResult::getValue() to get the validated valid value.
+- Added ValidResult::getValue() to get the validated valid value.
+- Added TransformerInterface to transform a value.
+- Added TransformerPipeline to sequentially transform a value with a pipeline of transformers.
 
 ### Changed